@@ -3,11 +3,12 @@ description: A new Flutter project.
 publish_to: none
 version: 1.0.0+1
 environment:
-  sdk: ">=2.12.0-259.8.beta <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  file_saver: ^0.0.10
   easy_web_view:
     path: ../
 
@@ -17,4 +18,6 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
+  assets:
+    - assets/
 
