@@ -1,6 +1,6 @@
 name: easy_web_view
 description: A super simple webview for Flutter Web and Mobile
-version: 1.5.0
+version: 2.0.0
 homepage: https://github.com/rodydavis/easy_web_view
 repository: https://github.com/rodydavis/plugins
 maintainer: Rody Davis (@rodydavis)
@@ -10,13 +10,12 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  webview_flutter: ^2.0.7
-  flutter_markdown: ^0.6.2
-  markdown: ^4.0.0
-  html2md: ^1.2.4
-  url_launcher: ^6.0.4
-  http: ^0.13.3
-#  permission_handler: ^5.0.1+1
+  flutter_markdown: ^0.6.9
+  html2md: ^1.2.5
+  http: ^0.13.4
+  markdown: ^4.0.1
+  url_launcher: ^6.0.17
+  webview_flutter: ^3.0.0
 
 dev_dependencies:
   flutter_test:
