@@ -1,4 +1,6 @@
 import 'package:easy_web_view/easy_web_view.dart';
+import 'package:example/ui/html_to_pdf/html_to_pdf.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -7,6 +9,9 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
+  EasyWebViewControllerWrapperBase? _controller;
+  EasyWebViewControllerWrapperBase? _controller2;
+  EasyWebViewControllerWrapperBase? _controller3;
   String src = 'https://flutter.dev';
   String src2 = 'https://flutter.dev/community';
   String src3 = 'http://www.youtube.com/embed/IyFZznAk69U';
@@ -51,6 +56,26 @@ class _HomeScreenState extends State<HomeScreen> {
                 },
               );
             }),
+            if ((_isHtml || _isMarkdown) && _controller != null)
+              IconButton(
+                icon: Icon(Icons.print),
+                onPressed: () async {
+                  final _c = _controller!;
+                  if (kIsWeb) {
+                    _c.postMessageWeb('print', '*');
+                  } else {
+                    // await _c.evaluateJSMobile(js);
+                  }
+                },
+              ),
+            IconButton(
+              onPressed: () => Navigator.of(context).push(
+                MaterialPageRoute(
+                  builder: (context) => HtmlToPdfTest(),
+                ),
+              ),
+              icon: Icon(Icons.picture_as_pdf),
+            ),
           ],
         ),
         body: _editing
@@ -147,7 +172,10 @@ class _HomeScreenState extends State<HomeScreen> {
                           flex: 1,
                           child: EasyWebView(
                             src: src,
-                            onLoaded: () {
+                            onLoaded: (controller) {
+                              setState(() {
+                                _controller = controller;
+                              });
                               print('$key: Loaded: $src');
                             },
                             isHtml: _isHtml,
@@ -160,10 +188,11 @@ class _HomeScreenState extends State<HomeScreen> {
                                 : WebNavigationDecision.navigate,
                             crossWindowEvents: [
                               CrossWindowEvent(
-                                  name: 'Test',
-                                  eventAction: (eventMessage) {
-                                    print('Event message: $eventMessage');
-                                  }),
+                                name: 'Test',
+                                eventAction: (eventMessage) {
+                                  // print('Event message: $eventMessage');
+                                },
+                              ),
                             ],
                             // width: 100,
                             // height: 100,
@@ -171,7 +200,10 @@ class _HomeScreenState extends State<HomeScreen> {
                       Expanded(
                         flex: 1,
                         child: EasyWebView(
-                          onLoaded: () {
+                          onLoaded: (controller) {
+                            setState(() {
+                              _controller2 = controller;
+                            });
                             print('$key2: Loaded: $src2');
                           },
                           src: src2,
@@ -201,7 +233,10 @@ class _HomeScreenState extends State<HomeScreen> {
                             width: (open) ? 500 : 0,
                             child: EasyWebView(
                                 src: src3,
-                                onLoaded: () {
+                                onLoaded: (controller) {
+                                  setState(() {
+                                    _controller3 = controller;
+                                  });
                                   print('$key3: Loaded: $src3');
                                 },
                                 isHtml: _isHtml,
@@ -224,6 +259,14 @@ class _HomeScreenState extends State<HomeScreen> {
 <html>
 <head>
 <title>Page Title</title>
+<script>
+window.addEventListener("message", (message) => {
+  console.log(message);
+  if (message.data === "print") {
+    window.print()
+  }
+})
+</script>
 </head>
 <body>
 <h1>This is a Heading</h1>
@@ -233,6 +276,14 @@ class _HomeScreenState extends State<HomeScreen> {
 """;
 
   String get markdownContent => """
+<script>
+window.addEventListener("message", (message) => {
+  console.log(message);
+  if (message.data === "print") {
+    window.print()
+  }
+})
+</script>
 # This is a heading
 ## Here's a smaller heading
 This is a paragraph
