@@ -1,4 +1,5 @@
 // ignore: avoid_web_libraries_in_flutter
+import 'dart:developer';
 import 'dart:html' as html;
 import 'dart:ui' as ui;
 
@@ -7,14 +8,38 @@ import 'package:flutter/material.dart';
 
 import 'impl.dart';
 
+class EasyWebViewControllerWrapper extends EasyWebViewControllerWrapperBase {
+  final html.IFrameElement _iframe;
+
+  EasyWebViewControllerWrapper._(this._iframe);
+
+  @override
+  Future<void> evaluateJSMobile(String js) {
+    throw UnsupportedError("the platform doesn't support this operation");
+  }
+
+  @override
+  Future<String> evaluateJSWithResMobile(String js) {
+    throw UnsupportedError("the platform doesn't support this operation");
+  }
+
+  @override
+  Object get nativeWrapper => _iframe;
+
+  @override
+  void postMessageWeb(dynamic message, String targetOrigin) =>
+      _iframe.contentWindow?.postMessage(message, targetOrigin);
+}
+
 class EasyWebView extends StatefulWidget implements EasyWebViewImpl {
   const EasyWebView({
     Key? key,
     required this.src,
-    required this.onLoaded,
+    this.onLoaded,
     this.height,
     this.width,
     this.webAllowFullScreen = true,
+    this.allow = "",
     this.isHtml = false,
     this.isMarkdown = false,
     this.convertToWidgets = false,
@@ -41,6 +66,9 @@ class EasyWebView extends StatefulWidget implements EasyWebViewImpl {
   final bool webAllowFullScreen;
 
   @override
+  final String allow;
+
+  @override
   final bool isMarkdown;
 
   @override
@@ -56,7 +84,7 @@ class EasyWebView extends StatefulWidget implements EasyWebViewImpl {
   final bool widgetsTextSelectable;
 
   @override
-  final void Function() onLoaded;
+  final OnLoaded? onLoaded;
 
   @override
   final List<CrossWindowEvent> crossWindowEvents;
@@ -66,16 +94,31 @@ class EasyWebView extends StatefulWidget implements EasyWebViewImpl {
 }
 
 class _EasyWebViewState extends State<EasyWebView> {
+  late Key effectiveKey;
+
+  // void testPrint(html.IFrameElement _iframe) {
+  //   _iframe.focus();
+  //   if (_iframe.srcdoc != null && _iframe.srcdoc!.isNotEmpty) {
+  //     final window = _iframe.contentWindow;
+  //     if (window != null) {
+  //       window.postMessage('print', '*');
+  //     }
+  //   }
+  //   log('iframe ready: ${_iframe.id}');
+  // }
+
   @override
   void initState() {
-    widget.onLoaded();
-    WidgetsBinding.instance?.addPostFrameCallback((timeStamp) {
-      final _iframe = _iframeElementMap[widget.key];
-      _iframe?.onLoad.listen((event) {
-        widget.onLoaded();
-      });
-    });
     super.initState();
+    effectiveKey = widget.key ?? ValueKey('');
+    WidgetsBinding.instance?.addPostFrameCallback((timeStamp) {
+      final _iframe = _iframeElementMap[effectiveKey];
+      if (_iframe != null) {
+        _iframe.onLoad.listen((event) {
+          widget.onLoaded?.call(EasyWebViewControllerWrapper._(_iframe));
+        });
+      }
+    });
   }
 
   @override
@@ -127,7 +170,7 @@ class _EasyWebViewState extends State<EasyWebView> {
         return AbsorbPointer(
           child: RepaintBoundary(
             child: HtmlElementView(
-              key: widget.key,
+              key: effectiveKey,
               viewType: 'iframe-$src',
             ),
           ),
@@ -139,19 +182,21 @@ class _EasyWebViewState extends State<EasyWebView> {
   static final _iframeElementMap = Map<Key, html.IFrameElement>();
 
   void _setup(String? src, double width, double height) {
-    final key = widget.key ?? ValueKey('');
     // ignore: undefined_prefixed_name
     ui.platformViewRegistry.registerViewFactory('iframe-$src', (int viewId) {
-      if (_iframeElementMap[key] == null) {
-        _iframeElementMap[key] = html.IFrameElement();
-      }
-      final element = _iframeElementMap[widget.key];
+      final element = _iframeElementMap[effectiveKey] ??= html.IFrameElement();
+      log('View Created: viewId: $viewId');
 
-      element!
-        ..style.border = '0'
-        ..allowFullscreen = widget.webAllowFullScreen
-        ..height = height.toInt().toString()
-        ..width = width.toInt().toString();
+      element
+            ..id = 'iframe-$viewId'
+            ..style.border = '0'
+            ..allowFullscreen = widget.webAllowFullScreen
+            ..allow = widget.allow
+            ..style.width = '100%'
+            ..style.height = '100%'
+          // ..height = height.toInt().toString()
+          // ..width = width.toInt().toString()
+          ;
 
       html.window.addEventListener('onbeforeunload', (event) async {
         final beforeUnloadEvent = (event as html.BeforeUnloadEvent);
@@ -179,17 +224,22 @@ class _EasyWebViewState extends State<EasyWebView> {
         });
       }
       String _src = src ?? '';
+      String? _srcDoc;
       if (src != null) {
         if (widget.isMarkdown) {
-          _src = "data:text/html;charset=utf-8," +
-              Uri.encodeComponent(EasyWebViewImpl.md2Html(src));
+          _srcDoc = EasyWebViewImpl.md2Html(src);
+          // _src = "data:text/html;charset=utf-8," + Uri.encodeComponent(_srcDoc);
         }
         if (widget.isHtml) {
-          _src = "data:text/html;charset=utf-8," +
-              Uri.encodeComponent(EasyWebViewImpl.wrapHtml(src));
+          _srcDoc = EasyWebViewImpl.wrapHtml(src);
+          // _src = "data:text/html;charset=utf-8," + Uri.encodeComponent(_srcDoc);
         }
       }
-      element..src = _src;
+
+      if (_srcDoc != null) {
+        element.srcdoc = _srcDoc;
+      }
+      element.src = _src;
       return element;
     });
   }
